requirements.txt
click==7.1.2
fastapi==0.61.2
h11==0.11.0
httptools==0.5.0
pydantic==1.7.2
starlette==0.13.6
uvicorn==0.12.3
uvloop==0.17.0
websockets==8.1
sentry-sdk==0.19.4
python-dotenv==0.15.0
sqlalchemy==1.3.20
SQLAlchemy-Continuum==1.3.11
python-multipart==0.0.5
aiofiles==0.6.0
psycopg2-binary==2.8.6
requests==2.25.0
python-jose[cryptography]==3.2.0
passlib[bcrypt]==1.7.4
alembic==1.4.3
password-strength==0.0.3.post2
pymongo==3.11.1
reportlab==3.6.12